requirements.txt
crewai==0.130.0
crewai-tools==0.45.0
langchain==0.3.25
langchain-community==0.3.24
langchain-groq==0.3.2
groq==0.23.1
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.170.0
amadeus==9.0.0
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3
requests==2.32.3
python-dateutil==2.9.0
streamlit==1.45.0
